Add files via upload
Player class is just a simple class that stores the disk color and name of each player, it can be extended upon for for keeping scores and turns in the future.
The PlayerMenu module is currently a work in progress, basically it is a popup window where each of the players can choose the game tags they have as well as chose they're disk color. My plan is to have this menu popup before the actual game starts, and communicate the colour and names of each player to the Player class where the values are stored.